services/chat-api/requirements.txt
# Core framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
supabase>=2.0.0

# HTTP client
httpx>=0.25.0

# Environment and utilities
python-dotenv>=1.0.0

# Authentication
PyJWT>=2.8.0
clerk-backend-api>=1.0.0

# Error tracking
sentry-sdk[fastapi]>=2.0.0

# AI/LLM (pre-staged for AMA-439 chat endpoints)
anthropic>=0.40.0
openai>=1.0.0

# TTS (ElevenLabs)
elevenlabs>=1.0.0

# SSE streaming
sse-starlette>=1.6.0

# KB content extraction (AMA-656)
readability-lxml>=0.8.1

# APNs push notifications
aioapns>=3.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pyyaml>=6.0.0,<7.0.0

# KB content extractors (AMA-657)
youtube-transcript-api>=0.6.0

# Load testing (optional, for tests/load/)
locust>=2.20.0,<3.0.0

# OpenTelemetry (AMA-506)
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
opentelemetry-exporter-otlp-proto-http>=1.22.0
opentelemetry-exporter-prometheus>=0.43b0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-instrumentation-httpx>=0.43b0
opentelemetry-instrumentation-logging>=0.43b0
opentelemetry-propagator-b3>=1.22.0

# spaCy NER (AMA-686)
spacy>=3.8.0
